// crypto-vault-expo/src/core/services/TonServices/type.ts
export interface Nftitem {
    address: string;
    index: number;
    owner: Owner;
    collection?: Collection;
    verified: boolean;
    metadata: Metadata;
    previews: Preview[];
    approvedBy?: any[];
    trust: TrustType;
    active?: boolean;
}

interface Preview {
    resolution: string;
    url: string;
}

interface Metadata {
    description: string;
    name: string;
    image: string;
    marketplace?: string;
    external_url?: string;
}

interface Collection {
    address: string;
    name: string;
    description: string;
}

interface Owner {
    address: string;
    is_scam?: boolean;
    is_wallet?: boolean;
}